pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "editor"
authors = [{name = "Alistair Buxton", email = "a.j.buxton@gmail.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
dynamic = ["version", "description"]
dependencies = [
    "PySide6",
    "pyqtconsole",
    "PyOpenGL",
    "csnake",
    "numpy",
]

[project.urls]
Source = "https://github.com/ali1234/racer"

[project.optional-dependencies]
test = [
    "pytest",
    "coverage",
]
